test: move root mocks under packages tests (#4588)
* Move execa mock under packages tests

* Remove better-sqlite3 mock

* Fix execa imports

* Remove execa mock
